rename(notificaiton): 파일 이름 변경 사항 반영 (#85)
